Watches are mostly done.
We still need to figure out how to attack a lastUpdatedDts referring to the last time the component was adjusted to each watch so that we can notify the user of updates.
Another TODO is to figure out how to notify the server of the event of a user viewing a component.