Add new Ext JS Version of Alerts Management
Users may add alerts or edit existing alerts.

Special attention should be paid to the functions which 'click'
the checkboxes while editing. These functions require that the checkbox
'name' attribute is the same as the 'name' that appears in the record
retrieved from the Store (and by extension the API).

At submission of the editAlertForm, attributes of the form record
are used to build a properly-formatted and nested JSON object to be
submitted to the API.