Validate form elements for Alerts Management
On submission of the form, the response from the server is received
and parsed into the proper format to use the validation controls
provided by Ext JS.

When checking for success after a PUT/POST form submission, the server
sends different types of responses depending on validation success. If
validation fails, it sends back an object with a property called
'success' which is set to false. If validation succeeds, it sends back
an object that does not contain this property. Therefore, somewhat
confusingly, if we see a 'success' property in the response, then we
know that the validation failed.

In order to allow for the Ext JS automated form validation markings,
we modify the 'id' of the email field, so that the markInvalid()
method can pick up the correct field to display the error message.