Enable activation status toggle for Alerts Mgmt
We add a method, actionToggleActivation(record), which takes a record
from the grid and toggles the active status of that Alert.

Ext seems to re-enable disabled buttons when the Store is reloaded, so
the Store must be reloaded, the grid completely deselected, and then
the buttons disabled. Without this order, Ext won't disable the buttons
as expected.