Enable form to add new Component Tags
We add a Window, tagAddWin, which has a form within, addTagForm, to
handle the addition of new tags for components. The form utilizes
two Stores:

- componentStore, which retrieves the list of components from which
the user will be able to select.

- securityTypeStore, which retrieves the codes and descriptions of the
available securityMarkingTypes. This store was introduced in babf69d and
is reused.

The form utilizes both Ext client-side validation and server-side
validation from the API.