Create initial framework for new Ext JS version of Component Questions
We set up a UI framework for the new component versions page, including
three panels across the screen in an hbox configuration. The idea is
that a user selects the component, selections a question belonging to
a component, and then the answers to to that question.

The three panels, componentPanel, questionPanel, and answerPanel, are
all inside of mainPanel, and are separated by splitters, which allow the
panels to be resized across the entire panel, adjusting automatically.