Load questions on component selection change
We attach the questionPanel to a store, questionStore, which in turn
has a proxy to call the API for questions associated with the given
componentId.

This proxy, questionStoreProxy, is created when a new function,
actionSelectedComponent, is invoked. The function is invoked when
a component is selected on the componentPanel grid. The questionStore is
then loaded with the correct questions.