Retrieve answers when question is selected
The actionSelectedQuestion(componentId, questionId) function is called
when a question on the questionGrid is selected. This function calls
the API to retrieve responses to the selected question.

The emptyText attribute of the answerPanel view includes some HTML
for an empty grid item -- something which happens automatically when
the emptyText is specified in the answerPanel's viewConfig upon
creation, but does not happen when the text is later set. This could be
an Ext JS bug or a failure to fully understand the grid/view paradigm.