Adjust filter comboboxes to filter on selection
Since the display of components is linked to the existence of questions
for that component, we choose to display the combobox for questions as
a selection to display components with either active or inactive
questions.

The "main" active status filter adjusts the componentStore's url to
make the proper request to get active or inactive status.

The answerPanel active status filter is a pure Ext filter and does not
make any additional requests. This is simply accomplished by adding a
.filter() call on the answerStore.

The main status filter is displayed as a toolbar above all three panels
and the refresh button below intentionally for aesthetical purposes.