Enable handlers for refresh and activation toggles
We add handler functions for the buttons which appear on the page,
the refresh, and the two activation toggles: question and answer.

These handlers perform validation to ensure that actual components,
questions, and answers are selected. Once the information is collected,
they call either the toggleQuestion or toggleAnswer method. These will
handle the proper API calls.

Another addition in the default value of 'A' in the answerPanel, to make
sure that the filter applies for Active status by default. This is
different from the questionPanel, which filters based on actual API
requests.