Use /view API method for questions panel
The questionPanel will only properly display the inactive questions
if we can request them specifically. In this case, we adjust the store's
API call to include the activeStatus as a parameter, and adjust to use
the questions/view path in order to due so, since the plain questions
path does not allow this filtering.

The user has a different name when requested when using this path, so
the username is adjusted.