Enable toggles of activation status
The two functions, toggleQuestion and toggleAnswer, which were
introduced in 64806a7, are filled out to perform the proper API
calls to ensure that both questions and answers can be activated
and deactivated.

When a question's status is toggled, the componentPanel's store must
be reloaded, just in case the toggling of the status resulted in that
component no longer being listed under the requested status.