Create page for Ext JS version of user profiles
The page includes a main userProfileGrid, with the necessary items as
found in the old page, as columns, along with a refresh button. This
will serve as a shell to improve the page into the full-service page.

Of special note is the necessity of using a special Reader on the proxy
for userProfileStore. This is because the API returns a list embedded
inside of an object under the name "data". This is different than usual,
and because this API call is likely used throughout the application, we
take this approach instead of attempting to modify the API.