Allow loadUser to be set on a parent window
When the user profileForm is being populated, we add another check to
see if a loadUser has been set on a parent profileWindow. If so, we
set the loadUser for the profileForm to the same as that of the parent.

This can be used when a page would like to use the OSF standard
userProfileWindow and set the User as a simple attribute of the window.