Use userProfileWindow to edit user profiles
The handler for the edit button, actionEditUser(), takes a record
and opens a userProfileWindow, specifying the user as found in the
record. Upon save, a callback function reloads the userProfileGrid
store.

The specification of the username uses the loadUser attribute on a
window, which was introduced in 8e7f7cd.