Switch from admin message to email based message
To match the previous user profile form, we send an e-mail to users
when the message button is pressed. This replaces the previous commit
which used the admin message capability which is a notification instead
of an email.

In addition, we set the messageWindow to be always on top, since it
will always be displaying on top of grids and such.