Disallow sorting by other columns
In the case of highlights, where there are very few entries, it should
not be necessary to sort the grid by any other column. If this were to
happen, the display position buttons would not appear to work to the
user, and would lead to unexpected results.

The position column is set to hidden since it is not needed to
ascertain the order, as the grid should already be ordered properly.

Makes progress on STORE-669.