Move highlights on grid on button press
When the up or down buttons are pressed, we move the record in the grid,
up or down, by inserting it at a different location. The new location
is determined by whether up or down is clicked.

In order to make buttons disable/enable at the correct times,
we fire the 'selectionchange' event on the highlight grid.

Makes progress on STORE-669.