Send API requests to properly reorder highlights
In order to properly reorder the list of highlights, each highlight
must have a unique position number. If this is not the case, the user
will experience unexpected issues when two or more highlights share
an ordering position.

To enforce this idea, the somewhat unfortunate best method is to change
the position of each highlight each time a single highlight is moved
up or down. That is, we must assign a number to each highlight in the
entire list each time there is a single change; every other path tends
to lend to unexpected results.

This is accomplished through setting each orderingPosition(the DB field)
equivalent to the grid row index position (as seen in Ext JS). When
combined with the changes made in e38b018, allows a proper reordering
which will not stray from expected conditions.

This commit should complete the client side of STORE-669.