Add status and scheduler control buttons
Allow the user to pause the scheduler. The button does not yet have
a handler, but the handler will send the API request to pause or
resume the scheduler, based on the scheduler state.

Acknowledging the issue in STORE-734, the scheduler state is displayed
directly adjacent to the control button, and is colored.