Enable actions for selected plugins
All actions, with the exception of actionDownloadPlugin, are simple API
calls based on the plugin ID. The success callback toasts the success
message and reloads the store.

The actionDownloadPlugin(record) function is based off of the
downloadCSV function found in the OSF CoreUtils. A link is generated
to the API download address and then clicked.