Show DB log status and toggle with single button
We show the status of the DB logger, which is captured in the values
returned by the sysConfigPropStore. This is accomplished by using
find() on the collection of data from the store given the code/key
'dblog.on'.

We use a function, updateDbLoggerStatus(reload) which updates the
label display and button given the status from sysConfigPropStore. If
reload is set to true, the store will be reloaded before the updates are
applied.

We also add a load hook on the sysConfigPropStore to update the label
and button whenever the store is loaded.