Disallow "hide on submission" flag on add
When the editAttributeWin window is being used for adding an attribute,
we disallow the "hide on submission" flag to be clicked. This is because
such a configuration requires a default code. We do not allow for a
default code on a "new" attribute, because codes must be created before
a default can be selected. Codes cannot be created until an attribute
exists for them. Therefore the only valid use of the "hide on submission"
flag will come after codes are created.