Save modified attributes on form submission
The formData must be compiled into an AttributeSaveType before it can
be submitted to the API. We also use the "useDataValues" parameter of
form.getValues() in order to ensure that we get "false" values for each
and every checkbox. This is because the API complains if it receives
null rather than "false" for the flags.