Populate attribute codes grid
The relevant columns for codes are added to the codesGrid, including
renderers on attribute code and highlight style. For the attribute code,
the attributeCodePk must be unpacked. For the highlight style, we apply
the 'alert-' prefix and set the result as a class in order to show the
color specified.