Pass record while adding attribute codes.
In previous cases, we did not require any information from a selected
item in order to "add" something. In this case however, we are managing
codes for a selected item, and to "add" a code, we need the information
associated with the attribute to which we are adding the code. Therefore
we adjust actionAddCode() to take a record, actionAddCode(record).