Adjust form to use attributeCodeView
We adjust some of the field names so that they work better
with the attributeCodeView API request instead of the attribtueCode API
request. This allows us to do filtering of inactive vs active.

We also must pass around the attributeType in order to complete the
request properly, so we have the properly selected attribute from the
attributeGrid.