Allow for multiple selection for exporting
We enable the checkbox selection model for attributeGrid, enabling
selections of multiple rows. Upon doing so, the standard edit buttons
will be disabled but the export button will remain available.

We also adjust actionExportAttribute to take multiple records from the
grid and actionImportAttribute to take no records, as it doesn't need
any to do its job.