Parse JSON into objects using mapper
Following the same style used in the ComponentStandardParser, we parse
the JSON found in the standard attribute file format and then use those
objects to send back to be persisted.

The GenericReader used simply takes the input stream and hands it over
to StringProcessor.defaultObjectMapper() to form a list of attributes.
As the nextRecord() method is called, this list gets returned one item
at a time, until empty.