Check for valid mimetype on import
As a form of format checking, we check to see if the mimetype of the
file matches what we are expecting. In the case of the SvcV-4 CSV
import, it is necessary that we check for a mimetype of ms-excel, since
Windows will often report this as a mimetype for a CSV file.