Revive old BaseAttributeParser
We revive the old BaseAttributeParser, which is used by the old
codebase for attribute importing. This class has been renamed
OldBaseAttributeParser, because BaseAttributeParser is now a name
being used by the new importing framework inheriting from AbstractParser.

This allows for imports to occur the old way or the new way.