Populate required entry types on edit
Upon edit, we populate the required entry types multiselector. In order
to do this, we take the restrictedTypes entry from the attribute
record, find the matching code in the searchStore of the multiselector,
and then add the matching searchStore record into the multiselector
store.

This allows us to use the the entry type description rather than
just the code, and is another layer of validation before the API request
is sent.