Add savedsearchlink plugin for TinyMCE
The plugin, savedsearchlink, is a plugin for TinyMCE, which uses
the Ext components found on openstorefront pages supporting the action
that will allow the user to selected a Saved Search. Upon selection
of the saved search, the user can click "Insert Link", and the plugin
will insert the link to the selected saved search.