Fix small bugs in searches page
Fixes two small issues issues with the searches page:

- An inability to save because Ext is unable to find the value of the
'searchName' textfield. This is remedied by doing a direct component
identification.

- Pressing the refresh button before selecting an active status caused
the refresh to load inactive items by default. To remedy, the initial
value of the active status selector should be 'A', ensuring that active
is the default selection (value) upon page load.