Retrieve and perform saved search
Upon time to populate the search results, we check for a savedSearchId
as given in the GET parameters. If set, we make an API request to
obtain the associated query parameters for the saved search. When they
are received, we de-seralize the objects, adjust them to to meet the
specifications of the doSearch() function and the API request, and call
the function to populate the search results.