Allow for listing of associated entry types
A multiselector, much like that used for restricted types, is used
in this case. If the multiselector is left empty, the attribute will
not be restricted to any set of entry types and will be permitted on
all entry types.

We open the multiselector by unchecking 'Allow for All Entry Types',
which we will leave as the default option.