Compile associated entry types for submission
Much like the required entry types list, we need to compile the list
into a format that the API would like to consume. Before we even assume
that we want to submit entry types, we check for the checkbox being
unchecked. Otherwise, we should not submit a list in order to allow all.