Only show attribute if associated with entry type
For the listing of optional attributes, we do not include them if
they have a list of associated component types and our submission is
not one of those component types.

If the attribute has no list of associated component types, it is
assumed to be an option for all entries, so it will always be included.

Makes progress on STORE-808.