Properly populate required attributes on edit
When we edit an attribute, we must make sure that the list of entry
types for which the attribute is required is populated properly. In this
case, we populate the grid with items from the search store. To do this,
we find the matching code in the search store and add it to the grid
store.

We will follow this same process for associatedComponentTypes, which
will be necessary for STORE-808.