Handle optional attributes required elsewhere
If the required flag is set, we still need to handle the case
where the entry type is _not_ one of the entry types in the requirements
list. This is accomplished by re-using the optional logic inside of the
required case.

Since this involves a fair amount of nested if statements, we add some
comments to explain what's going on.

This should finish STORE-808.