STORE-805 Calendar positioning in scroll
When the datepicker is selected, followed by moving of the form in
which it is contained, opening the datepicker again results in a
"blank" screen, manifested presumably by a location/positioning issue.

We add a CSS directive to modify the position approach for the browser.
One downside is that the widget itself can display in undesirable
locations, but it is much better than the alternative. A better fix
may come around in Sencha's fix.