Create API endpoint for user profile export
In pursuit of user profile export as detailed in STORE-793, we add
an API endpoint for doing so. The API will take userIds as FormParams,
to be used to compile the proper JSON objects for each user profile.

In this commit, we test for proper fucntion of the endpoint with a
sample response.