Adjust userProfiles to be UserProfileView type
For some reason, the standard userProfile can't be serialized due to
a cycle in the object -- doesn't really make any sense since all of
the members objects are Strings. Seems to be something related to
orientdb. Using UserProfileView seems to fix the issue, although we'll
need to potentially create another view to handle all user information
if the view isn't sufficient.

STORE-793