Allow for users to be selected on main profile grid
We allow a set of initialBccUsers to be specified upon creation of the
message window. Also, in order to populate the field in the expected
format, we create a string joining the arrays with a semicolon, since
the default is just a comma.

Additionally, we ask the user if they want to populate the 'To' field
or 'BCC' field.

This completes STORE-941.