Create Temporary Media entity and resource
The TemporaryMedia entity will be used to store media that the server
is temporarily storing for the client's use.

In this case, Temporary
Media is introduced primarily for the purposes of achieving STORE-781,
in which the client captures pasted image URLs and asks the server to
download them for local storage. Upon saving, these saved images are
then modified to become component media.