STORE-871 Handling of Inline Media
In an incredible display of either foolishness on the behalf of the
developer, or foolishness on the behalf of JavaScript's design
decisions is exhibited here. The replace function for a string only
replaces the first occurrence of a string, therefore the multitude
of unexplained issues is suddenly making sense.

In this case, we create a global RegExp to replace all of the occurences
and magically don't encounter trouble.

The moral of this story is: while debugging, don't think that some
library or base javascript function does what you think it does.

STORE-871