Adjust submission panel to allow user-generated attribute codes
Upon an attempt to submit, the code is checked to see if the associated
type allows user-generated codes. If so, a check is performed to see
if the code is an already valid code. If not, a request to generate
a new attribute code is sent and the new attribute code is replaced in
the form before finally submitting it to be saved.

STORE-1011