Adjust inline handler for HTML entities
TinyMCE's getContent() function is a strange one in that it tries
to clean up the HTML before it sends it over. That's all nice except
strings with HTML encoded entities are different than strings without.
That breaks everything.

So, we work around this by using the browser to re-decode the html so
we can actually do our text replacement how you'd expect.