Update ElasticSearchManager.java
Changed auto created log to use the class logger for consistency.
Move the flag indexCreated to avoid calling the  create over and over if the index was already created,  The only time it will repeat is in the case where it couldn't verify with elasticsearch do to communication failure.  In which case it should try again once the communication failure is resolved.